Remove need to include router module
